feat: Implement pagination for pharmacy orders
This commit introduces pagination to the `PharmacyOrdersScreen` to improve performance and user experience when loading a large number of orders.

Key changes:

- **`GetPharmacyOrdersUseCase`:** Updated to accept `page` and `pageSize` parameters.
- **`RepositoryImpl`:** `getPharmacyOrders` now passes `page` and `pageSize` to the `remoteDataSource`.
- **`PharmacyOrdersViewModel`:**
    - Added `currentPage`, `pageSize`, `isLastPage`, and `isLoading` properties to manage pagination state.
    - Implemented `loadNextPage()` to fetch subsequent pages of orders.
    - `refreshOrders()` now resets pagination and fetches the first page.
    - Orders are now appended to a local `orders` list.
    - `getPharmacyOrders()` is renamed to `fetchPharmacyOrders()` and updated to handle pagination logic.
- **`PharmacyOrdersScreen`:**
    - `getPharmacyOrders()` is replaced with `refreshOrders()` in `LaunchedEffect`.
    - A `LaunchedEffect` is added within the `LazyColumn` to trigger `loadNextPage()` when the user scrolls near the end of the current list.
    - When an error occurs while loading order details in the bottom sheet, an `ErrorView` is displayed with a retry action.

The `pageSize` is set to 15.

+++ app/src/main/java/com/example/atonce_admin/data/remote/dto/OrdersStatusResponse.kt
@@ -9,7 +9,7 @@ data class OrderStatusResponse(
 )
 
 data class Item(
-    val deliveredRevenue: Int,
+    val deliveredRevenue: Double,
     val orders: List<Order>,
     val ordersCount: Int,
     val totalPrice: Double,
